# Optimize typechecking