--- a/apps/builder/package.json
+++ b/apps/builder/package.json
@@ -119,7 +119,7 @@
     "jest": "^29.6.1",
     "react-test-renderer": "^18.2.0",
     "type-fest": "^3.7.1",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   },
   "sideEffects": false,
   "license": "MIT"