--- a/package.json
+++ b/package.json
@@ -35,7 +35,7 @@
     "storybook": "^7.0.26",
     "tsx": "^3.12.6",
     "turbo": "^1.8.6",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   },
   "engines": {
     "node": "18",