--- a/packages/asset-uploader/package.json
+++ b/packages/asset-uploader/package.json
@@ -32,7 +32,7 @@
     "@webstudio-is/scripts": "workspace:^",
     "@webstudio-is/tsconfig": "workspace:^",
     "jest": "^29.6.1",
-    "typescript": "5.1.3",
+    "typescript": "5.1.6",
     "zod": "^3.21.4"
   },
   "exports": {