--- a/packages/authorization-token/package.json
+++ b/packages/authorization-token/package.json
@@ -20,7 +20,7 @@
   "devDependencies": {
     "@webstudio-is/scripts": "workspace:^",
     "@webstudio-is/tsconfig": "workspace:^",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   },
   "exports": {
     ".": {