--- a/packages/css-engine/package.json
+++ b/packages/css-engine/package.json
@@ -6,7 +6,7 @@
   "homepage": "https://webstudio.is",
   "type": "module",
   "scripts": {
-    "typecheck": "tsc --noEmit --emitDeclarationOnly false",
+    "typecheck": "tsc",
     "checks": "pnpm typecheck && pnpm test",
     "dev": "build-package --watch",
     "build": "build-package",