--- a/packages/css-vars/package.json
+++ b/packages/css-vars/package.json
@@ -15,7 +15,7 @@
   "devDependencies": {
     "@webstudio-is/scripts": "workspace:^",
     "@webstudio-is/tsconfig": "workspace:^",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   },
   "exports": {
     "source": "./src/index.ts",