--- a/packages/dashboard/package.json
+++ b/packages/dashboard/package.json
@@ -21,7 +21,7 @@
   "devDependencies": {
     "@webstudio-is/scripts": "workspace:^",
     "@webstudio-is/tsconfig": "workspace:^",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   },
   "exports": {
     ".": {