--- a/packages/design-system/package.json
+++ b/packages/design-system/package.json
@@ -30,7 +30,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "tsx": "^3.12.6",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   },
   "peerDependencies": {
     "react": "^18.2.0",