--- a/packages/fonts/package.json
+++ b/packages/fonts/package.json
@@ -24,7 +24,7 @@
     "@webstudio-is/scripts": "workspace:^",
     "@webstudio-is/tsconfig": "workspace:^",
     "jest": "^29.6.1",
-    "typescript": "5.1.3",
+    "typescript": "5.1.6",
     "zod": "^3.21.4"
   },
   "peerDependencies": {