--- a/packages/form-handlers-mailchannels-test/package.json
+++ b/packages/form-handlers-mailchannels-test/package.json
@@ -12,7 +12,7 @@
   "devDependencies": {
     "@cloudflare/workers-types": "^3.0.0",
     "@webstudio-is/tsconfig": "workspace:^",
-    "typescript": "5.1.3",
+    "typescript": "5.1.6",
     "wrangler": "^2.18.0"
   }
 }