--- a/packages/form-handlers/package.json
+++ b/packages/form-handlers/package.json
@@ -18,7 +18,7 @@
   "devDependencies": {
     "@webstudio-is/scripts": "workspace:^",
     "@webstudio-is/tsconfig": "workspace:^",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   },
   "exports": {
     "source": "./src/index.ts",