--- a/packages/generate-arg-types/package.json
+++ b/packages/generate-arg-types/package.json
@@ -21,7 +21,7 @@
     "@webstudio-is/scripts": "workspace:^",
     "@webstudio-is/tsconfig": "workspace:^",
     "tsx": "^3.12.6",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   },
   "files": [
     "lib/*",