--- a/packages/http-client/package.json
+++ b/packages/http-client/package.json
@@ -21,7 +21,7 @@
     "@webstudio-is/scripts": "workspace:^",
     "@webstudio-is/tsconfig": "workspace:^",
     "jest": "^29.6.1",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   },
   "exports": {
     "source": "./src/index.ts",