--- a/packages/image/package.json
+++ b/packages/image/package.json
@@ -30,7 +30,7 @@
     "jest": "^29.6.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   },
   "peerDependencies": {
     "react": "^18.2.0",