--- a/packages/prisma-client/package.json
+++ b/packages/prisma-client/package.json
@@ -23,7 +23,7 @@
     "@webstudio-is/tsconfig": "workspace:^",
     "prisma": "^4.14.0",
     "tsx": "^3.12.6",
-    "typescript": "5.1.3",
+    "typescript": "5.1.6",
     "zod": "^3.21.4"
   },
   "exports": {