--- a/packages/project-build/package.json
+++ b/packages/project-build/package.json
@@ -26,7 +26,7 @@
   ],
   "sideEffects": false,
   "scripts": {
-    "typecheck": "tsc --noEmit --emitDeclarationOnly false",
+    "typecheck": "tsc",
     "test": "NODE_OPTIONS=--experimental-vm-modules jest",
     "checks": "pnpm typecheck && pnpm test",
     "dev": "build-package --watch",