--- a/packages/project-build/package.json
+++ b/packages/project-build/package.json
@@ -47,6 +47,6 @@
     "@webstudio-is/tsconfig": "workspace:^",
     "jest": "^29.6.1",
     "type-fest": "^3.7.1",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   }
 }