--- a/packages/react-sdk/package.json
+++ b/packages/react-sdk/package.json
@@ -26,7 +26,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "type-fest": "^3.7.1",
-    "typescript": "5.1.3",
+    "typescript": "5.1.6",
     "zod": "^3.21.4"
   },
   "peerDependencies": {