--- a/packages/scripts/package.json
+++ b/packages/scripts/package.json
@@ -21,6 +21,6 @@
     "@types/node": "^18.11.18",
     "@webstudio-is/tsconfig": "workspace:^",
     "tsx": "^3.12.6",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   }
 }