--- a/packages/sdk-components-react/package.json
+++ b/packages/sdk-components-react/package.json
@@ -65,6 +65,6 @@
     "@webstudio-is/tsconfig": "workspace:^",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   }
 }