--- a/packages/trpc-interface/package.json
+++ b/packages/trpc-interface/package.json
@@ -26,7 +26,7 @@
     "@webstudio-is/jest-config": "workspace:^",
     "@webstudio-is/scripts": "workspace:^",
     "@webstudio-is/tsconfig": "workspace:^",
-    "typescript": "5.1.3"
+    "typescript": "5.1.6"
   },
   "exports": {
     "./index.server": {